technical<service book test> add a dummy BookService test add a Dockerfile

+++ entity/src/main/java/org/thibaut/thelibrary/entity/BookEntity.java
@@ -31,7 +31,7 @@ public class BookEntity {
 
 	private String language;
 
-	private Integer isbn;
+	private Long isbn;
 
 	private DateTime publicationDate;
